Update README.md
You've done excellent work on this addon! I just had a minor documentation suggestion. As a newbie, I struggled for an hour before I realized I was referencing the wrong entities when trying to do playback. On HASS OS 2022.8 and a Symfonix speaker, audio playback only resumes if I target the new Sonos entities :) This wasn't obvious to me at the beginning.